NAAF-Research-Agents/requirements.txt
google-adk==1.24.1
httpx==0.28.1
pydantic==2.11.1
python-dotenv==1.1.0
